Add check for valid feature_idx in save_feature_centric_vis
Added an assert which ensures that the `feature_idx` specified a user is present in the `feature_data_dict`. If this is not the case then the html will still be produced but will not contain divs for the other features, which made the source of the error a bit tricky to track down.